Make sure the next scheduled time is in the future
If a query takes longer than its interval to finish, or it is delayed in
starting due to scheduling contention, schedule_time + interval can be in the
past, which will effectively result in the query running all the time.

This fix skips potential run points that have already occurred to prevent this.