Periods aren't valid in metric names, so replace with underscores
Metric names generated from ES results may have periods in them (e.g. the
percentiles aggregation). This solution also silently replaces periods in
user-selected parts of metric names, but that's not a disastrous
side-effect for a much simpler implementation.

There will be other characters that are invalid in metric names, but lets
ignore them unless we find a ES query that generates them.