Convert cluster metrics to custom collectors
Queries to the `_cluster/health`, `_nodes/stats`, and `_stats` endpoints of
the Elasticsearch cluster will now be triggered when the exporter's
metric endpoint is called, rather than on a timer. This effectively allows
the upstream Prometheus server to set query interval for these metrics with
its scrape interval. This is the recommended approach for instrumenting
external systems.

Note that this approach isn't followed for the custom data queries, as
these are more likely to have varying costs, producing a variance in ideal
frequency across the different queries, and increasing the potential
impact of an upstream server naively scraping too frequently.