Fix issue where main config file is required even when queries are disabled
The main config file shouldn't be required if --query-disable is set.
While it isn't used in this case, Click was trying to open the default file
anyway when checking the CLI options, resulting in an error.

Now Click won't error if it can't find the main config file. If queries
aren't disabled, and no config file exists or includes any metrics, the
exporter will log an error and exit.