Fix nodes stats parsing for AWS managed Elasticsearch clusters
AWS doesn't include the path of data directories in the `_nodes/stats`
output. This caused an error when the parser tried to extract the path to
use as a label.

When parsing a list of buckets (like the list of data directories), the
parser will now fall back to using bucket's position in the list as the
bucket name if the expected bucket name key isn't present.